+++ src/components/smart-contracts/CreateNewFund.js
@@ -1,6 +1,6 @@
-import { ContractFactory } from 'ethers';
 import * as ethers from 'ethers';
 import contractData from './contracts/CollectiveInvestmentSchemeV2.json';
+import { loadSigner } from './loadFund';
 
 function solDateConversion(date) {
   if (date instanceof Date) {
@@ -15,25 +15,7 @@ function solDateConversion(date) {
 
 const createNewFund = async ({fundName, underlyingAsset, offerClosingTime, orderExpiration, maturity}) => {
     try {
-      if (typeof window.ethereum === 'undefined') {
-        throw new Error('MetaMask is not installed...');
-      }
-      const accounts = await window.ethereum.request({ method: 'eth_requestAccounts' });
-      if (accounts.length == 0) {
-        throw new Error('Fail to get account.');
-      }  
-      {
-        // prohibit mainnet
-        const hexChainId = await window.ethereum.request({ method: 'eth_chainId' });
-        const chainId = parseInt(hexChainId);
-        console.log(`chainId=${chainId}`)
-        if (chainId != 5) {
-          throw new Error('Only Goerli TestNet(chainId=5) is available.');
-        }
-      }
-      console.log(accounts[0]);
-      const provider = new ethers.providers.Web3Provider(window.ethereum);
-      const signer = provider.getSigner();
+      const signer = await loadSigner();
       console.log("Account:", await signer.getAddress());
       const factory = new ethers.ContractFactory(contractData.abi, contractData.bytecode, signer);
